Review
- move qsort to a submodule of stdlib
- add more tests
